requirements.txt
Flask==1.0.2
kafka-python==1.4.3
ipython==6.4.0
